feat(tasks): Move step executor to recording namespace
